refactor: LottoResult 클래스 네이밍 변경, compareTo 삭제, toString 추가
